feat: Add sha256 allowlist matching to allow only very precise file or script contents
This adds sha256 hashes to all findings with file or script contents. Hence, no additional time is spent on reading files, only the in RAM calculation on actual findings.
The allowlist schema is extended with a sha256 match property and therefore allows only allowing very precise content payloads to pass the checks.